Adding all new Admin Service for Admin Managment & Docker Compose support for DB's
Within this commit a new service is included that allows the dynamic registration of OAuth2.0 clients, it also includes the management of sessions with persistence in Redis and lastly, support for Docker is included and in this way to be able to port the Databases. project data directly to the code in order to obtain more portability.